Main

+++ Lesson_2/ArithmeticMean/Program.cs
@@ -1,4 +1,4 @@
-﻿int a = int.Parse(Console.ReadLine());
-int b = int.Parse(Console.ReadLine());
-Console.WriteLine((a+b)/2);
+﻿float a = int.Parse(Console.ReadLine());
+float b = int.Parse(Console.ReadLine());
+Console.WriteLine((float)(a+b)/2);
 Console.ReadKey();
